+++ source/core/parameters.cc
@@ -78,11 +78,13 @@ namespace Parameters
   {
     prm.enter_subsection("simulation control");
     {
-      prm.declare_entry("method",
-                        "steady",
-                        Patterns::Selection("steady|steady_bdf|bdf1|bdf2|bdf3"),
-                        "The kind of solver for the linear system. "
-                        "Choices are <steady|steady_bdf|bdf1|bdf2|bdf3>.");
+      prm.declare_entry(
+        "method",
+        "steady",
+        Patterns::Selection(
+          "steady|steady_bdf|bdf1|bdf2|bdf3|sdirk22|sdirk33|sdirk43"),
+        "The time integration scheme. "
+        "Choices are <steady|steady_bdf|bdf1|bdf2|bdf3|sdirk22|sdirk33|sdirk43>.");
 
       prm.declare_entry(
         "bdf startup method",
@@ -215,6 +217,12 @@ namespace Parameters
         method = TimeSteppingMethod::bdf2;
       else if (sv == "bdf3")
         method = TimeSteppingMethod::bdf3;
+      else if (sv == "sdirk22")
+        method = TimeSteppingMethod::sdirk22;
+      else if (sv == "sdirk33")
+        method = TimeSteppingMethod::sdirk33;
+      else if (sv == "sdirk43")
+        method = TimeSteppingMethod::sdirk43;
       else
         {
           std::runtime_error("Invalid time stepping scheme");