+++ source/solvers/heat_transfer.cc
@@ -275,6 +275,10 @@ template <int dim>
 void
 HeatTransfer<dim>::setup_assemblers()
 {
+  AssertThrow(
+    is_sdirk(this->simulation_control->get_assembly_method()) == false,
+    ExcMessage("The SDIRK scheme is not yet supported for this physics"));
+
   this->assemblers.clear();
 
   // Laser heat source