+++ tests/core/sdirk_stage.cc
@@ -12,6 +12,10 @@ test()
   // 12 digits of precision for the output
   deallog << std::setprecision(12) << std::scientific;
 
+  deallog
+    << "This test returns the values of the Butcher's table associated with each SDIRK method"
+    << std::endl;
+
   deallog << "Testing SDIRK22 coefficients" << std::endl;
 
   // Important note : the nomenclature used for the name of the SDIRK methods
@@ -30,6 +34,7 @@ test()
       deallog << "  a_ij: ";
       for (const auto &a : data.a_ij)
         deallog << std::setw(12) << a << " ";
+      deallog << "\n  b_i : " << data.b_i << std::endl;
       deallog << "\n  c_i : " << data.c_i << std::endl;
     }
 
@@ -50,6 +55,28 @@ test()
       deallog << "  a_ij: ";
       for (const auto &a : data.a_ij)
         deallog << std::setw(12) << a << " ";
+      deallog << "\n  b_i : " << data.b_i << std::endl;
+      deallog << "\n  c_i : " << data.c_i << std::endl;
+    }
+
+  deallog << "<----------------------------->\n" << std::endl;
+
+  deallog << "Testing SDIRK43 coefficients" << std::endl;
+
+  SDIRKTable table43 =
+    sdirk_table(Parameters::SimulationControl::TimeSteppingMethod::sdirk43);
+  const unsigned int n_stages43 = table43.A.m();
+
+  // Data printed at each stage
+  for (unsigned int stage_i = 1; stage_i <= n_stages43; ++stage_i)
+    {
+      SDIRKStageData data(table43, stage_i);
+
+      deallog << "\nStage " << stage_i << ":" << std::endl;
+      deallog << "  a_ij: ";
+      for (const auto &a : data.a_ij)
+        deallog << std::setw(12) << a << " ";
+      deallog << "\n  b_i : " << data.b_i << std::endl;
       deallog << "\n  c_i : " << data.c_i << std::endl;
     }
 }